Remove Connecting Intents
A connecting intent had direct access to the MessagePort connecting to the peer,
and was not restricted to the offer/accept/reply flow. However, PostingOffers can
include MessagePorts in offer messages, so in theory there isn't anything you can
do with a connecting intent that you can't with a posting intent. In the interest
of keeping the API as small as possible, this removes connecting intents.